refactor: 구독 요청 경로 변경
* 기능을 ParentController와 ParentService에서 PickUpController와 PickUpService로 이동
