+++ BE/src/main/java/ifive/idrop/pickup/controller/PickUpController.java
@@ -3,6 +3,8 @@ package ifive.idrop.pickup.controller;
 import ifive.idrop.auth.resolver.Login;
 import ifive.idrop.common.dto.BaseResponse;
 import ifive.idrop.driver.domain.Driver;
+import ifive.idrop.parent.domain.Parent;
+import ifive.idrop.parent.dto.SubscriptionRequest;
 import ifive.idrop.pickup.domain.PickUpHistory;
 import ifive.idrop.common.exception.BusinessException;
 import ifive.idrop.common.exception.ErrorCode;
@@ -11,6 +13,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -21,8 +24,13 @@ import java.util.concurrent.ExecutionException;
 @RequiredArgsConstructor
 @RestController
 public class PickUpController {
-
     private final PickUpService pickUpService;
+
+    @PostMapping("/subscriptions")
+    public BaseResponse<String> postSubscription(@Login Parent parent, @RequestBody SubscriptionRequest request) {
+        pickUpService.subscribe(parent, request);
+        return BaseResponse.success();
+    }
 
     @PostMapping("/driver/pickup")
     public BaseResponse<String> startOrEndPickUp(@Login Driver driver, Long childId, @ModelAttribute MultipartFile image, String message) throws ExecutionException, InterruptedException {
