+++ BE/src/main/java/ifive/idrop/pickup/repository/PickUpRepository.java
@@ -44,10 +44,6 @@ public class PickUpRepository {
         return query.getResultList();
     }
 
-    public void savePickUpLocation(PickUpLocation location) {
-        em.persist(location);
-    }
-
     public void savePickUpInfo(Subscription subscription) {
         em.persist(subscription);
     }
